+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/jss/compare/v1.2.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Adds `RESERVED_USERNAMES` to config
 
 [1.2.0](https://github.com/jshwi/jss/releases/tag/v1.2.0) - 2021-10-11
 ------------------------------------------------------------------------